+++ commands/transfer_funds_test.go
@@ -1061,6 +1061,35 @@ func TestTransferFunds(t *testing.T) {
 		},
 		{
 			transfer: commandspb.Transfer{
+				FromAccountType: vega.AccountType_ACCOUNT_TYPE_GENERAL,
+				ToAccountType:   vega.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION,
+				Kind: &commandspb.Transfer_Recurring{
+					Recurring: &commandspb.RecurringTransfer{
+						StartEpoch: 10,
+						EndEpoch:   ptr.From(uint64(11)),
+						Factor:     "1",
+						DispatchStrategy: &vega.DispatchStrategy{
+							AssetForMetric:       "080538b7cc2249de568cb4272a17f4d5e0b0a69a1a240acbf5119d816178daff",
+							Metric:               vega.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
+							DistributionStrategy: vega.DistributionStrategy_DISTRIBUTION_STRATEGY_PRO_RATA,
+							EntityScope:          vega.EntityScope_ENTITY_SCOPE_INDIVIDUALS,
+							IndividualScope:      vega.IndividualScope_INDIVIDUAL_SCOPE_IN_TEAM,
+							TeamScope:            []string{"zohafr"},
+							StakingRequirement:   "1",
+							NotionalTimeWeightedAveragePositionRequirement: "2",
+							WindowLength: 1,
+						},
+					},
+				},
+				To:        "84e2b15102a8d6c1c6b4bdf40af8a0dc21b040eaaa1c94cd10d17604b75fdc35",
+				Asset:     "080538b7cc2249de568cb4272a17f4d5e0b0a69a1a240acbf5119d816178daff",
+				Amount:    "1",
+				Reference: "testing",
+			},
+			errString: "transfer.kind.dispatch_strategy.team_scope (should not be set when entity_scope is set to ENTITY_SCOPE_INDIVIDUALS)",
+		},
+		{
+			transfer: commandspb.Transfer{
 				FromAccountType: vega.AccountType_ACCOUNT_TYPE_VESTED_REWARDS,
 				ToAccountType:   vega.AccountType_ACCOUNT_TYPE_MARGIN,
 				Kind: &commandspb.Transfer_Recurring{